scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
supabase>=2.0.0
python-slugify>=8.0.0
anthropic>=0.40.0
